pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "freshai"
version = "0.1.0"
description = "AI Assistant for Crime Investigators using LLM and VLM capabilities"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "FreshAI Team", email = "team@freshai.com"},
]
keywords = ["ai", "crime investigation", "llm", "vlm", "ollama", "transformers"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Legal Industry",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.28.0",
    "aiohttp>=3.8.0",
    "pydantic>=2.0.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "pillow>=9.0.0",
    "numpy>=1.21.0",
    "opencv-python>=4.7.0",
    "python-dotenv>=1.0.0",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "asyncio>=3.4.3",
    "matplotlib>=3.5.0",
    "httpx>=0.28.1",
    "bitsandbytes>=0.47.0",
    "accelerate>=1.10.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.0.280",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/Michael89/FreshAI"
Repository = "https://github.com/Michael89/FreshAI"

[project.scripts]
freshai = "freshai.cli:main"

# Package discovery handled by hatchling automatically

[tool.black]
line-length = 88
target-version = ['py39']

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.0.280",
    "mypy>=1.0.0",
]

[tool.uv.sources]
# Add any local development sources here if needed